+++ RouteComposer/Classes/Assemblies/StepAssembly.swift
@@ -211,87 +211,3 @@ public extension StepAssembly where FC: ContainerFactory & NilEntity {
     }
 
 }
-
-// MARK: - Shorthand overloads to enable `.using(.)`
-
-public extension StepAssembly where FC: Factory {
-    /// Enables shorthand `.using(.present(...))` by providing a concrete expected type.
-    final func using(_ action: ViewControllerActions.PresentModallyAction) -> StepChainAssembly<ViewController, Context> {
-        usingAction(action)
-    }
-
-    /// Enables shorthand `.using(.replaceRoot(...))`
-    final func using(_ action: ViewControllerActions.ReplaceRootAction) -> StepChainAssembly<ViewController, Context> {
-        usingAction(action)
-    }
-
-    /// Enables shorthand `.using(.nilAction)`
-    final func using(_ action: ViewControllerActions.NilAction) -> StepChainAssembly<ViewController, Context> {
-        usingAction(action)
-    }
-}
-
-public extension StepAssembly where FC: ContainerFactory {
-    /// Enables shorthand `.using(.presentModally(...))` by providing a concrete expected type.
-    final func using(_ action: ViewControllerActions.PresentModallyAction) -> StepChainAssembly<ViewController, Context> {
-        usingAction(action)
-    }
-
-    /// Enables shorthand `.using(.replaceRoot(...))`
-    final func using(_ action: ViewControllerActions.ReplaceRootAction) -> StepChainAssembly<ViewController, Context> {
-        usingAction(action)
-    }
-
-    /// Enables shorthand `.using(.nilAction)`
-    final func using(_ action: ViewControllerActions.NilAction) -> StepChainAssembly<ViewController, Context> {
-        usingAction(action)
-    }
-}
-
-public extension StepAssembly where FC: Factory {
-    /// Enables shorthand `.using(.push)` by providing a concrete expected type.
-    final func using(_ action: NavigationControllerActions.PushAction<UINavigationController>) -> ContainerStepChainAssembly<UINavigationController, ViewController, Context> {
-        usingAction(action)
-    }
-
-    /// Enables shorthand `.using(.pushAsRoot)`
-    final func using(_ action: NavigationControllerActions.PushAsRootAction<UINavigationController>) -> ContainerStepChainAssembly<UINavigationController, ViewController, Context> {
-        usingAction(action)
-    }
-
-    /// Enables shorthand `.using(.pushReplacingLast)`
-    final func using(_ action: NavigationControllerActions.PushReplacingLastAction<UINavigationController>) -> ContainerStepChainAssembly<UINavigationController, ViewController, Context> {
-        usingAction(action)
-    }
-}
-
-public extension StepAssembly where FC: ContainerFactory {
-    /// Enables shorthand `.using(.push)` by providing a concrete expected type.
-    final func using(_ action: NavigationControllerActions.PushAction<UINavigationController>) -> ContainerStepChainAssembly<UINavigationController, ViewController, Context> {
-        usingAction(action)
-    }
-
-    /// Enables shorthand `.using(.pushAsRoot)`
-    final func using(_ action: NavigationControllerActions.PushAsRootAction<UINavigationController>) -> ContainerStepChainAssembly<UINavigationController, ViewController, Context> {
-        usingAction(action)
-    }
-
-    /// Enables shorthand `.using(.pushReplacingLast)`
-    final func using(_ action: NavigationControllerActions.PushReplacingLastAction<UINavigationController>) -> ContainerStepChainAssembly<UINavigationController, ViewController, Context> {
-        usingAction(action)
-    }
-}
-
-public extension StepAssembly where FC: Factory {
-    /// Enables shorthand `.using(.addTab)` by providing a concrete expected type.
-    final func using(_ action: TabBarControllerActions.AddTabAction<UITabBarController>) -> ContainerStepChainAssembly<UITabBarController, ViewController, Context> {
-        usingAction(action)
-    }
-}
-
-public extension StepAssembly where FC: ContainerFactory {
-    /// Enables shorthand `.using(.addTab)` by providing a concrete expected type.
-    final func using(_ action: TabBarControllerActions.AddTabAction<UITabBarController>) -> ContainerStepChainAssembly<UITabBarController, ViewController, Context> {
-        usingAction(action)
-  }
-}
